migration part2

+++ tyama_live3051-PKU2006-HNU10343_golf.c
@@ -1,2 +1,2 @@
 // Solve Ka = (mx)(nx) / (A-x) <=> mnx2 - Kax + KaA = 0 and print log10(1/mx)
-main(m,n){float k,a;for(;scanf("%f%f%d%d",&k,&a,&m,&n),k;printf("%.3f\n",log10(n*2/(sqrt(k*k+4*m*n*k*a)-k))));}
+main(m,n){double k,a;for(;scanf("%lf%lf%d%d",&k,&a,&m,&n),k;printf("%.3f\n",log10(n*2/(sqrt(k*k+4*m*n*k*a)-k))));}
